test: merge countdown timer specs
